renamed routes in routes to router to avoid confusion
